+++ ext/nokogiri/xml_xpath_context.c
@@ -46,7 +46,7 @@ static void ruby_funcall(xmlXPathParserContextPtr ctx, int nargs)
 
   xpath_handler = (VALUE)(ctx->context->userData);
 
-  argv = (VALUE *)calloc((unsigned int)nargs, sizeof(VALUE));
+  argv = (VALUE *)calloc((size_t)nargs, sizeof(VALUE));
   doc = DOC_RUBY_OBJECT(ctx->context->doc);
 
   i = nargs - 1;
